crawler/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
html5lib>=1.1
PyJWT>=2.8.0